server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.10.0
python-dotenv>=1.0.0
aiofiles>=24.1.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.23
databricks-sdk>=0.18.0 